refactor(type): add IDiscordMessageFlag type
since this type is only an alias for a string the only purpose here is to distinguish a bit more a simple message from a reduce, flag message
