+++ src/features/discord/messages/classes/commands/flags/discord-command-flag.ts
@@ -2,6 +2,7 @@ import _ from "lodash";
 import { DiscordCommandFlagTypeEnum } from "../../../enums/commands/discord-command-flag-type.enum";
 import { IDiscordCommandFlag } from "../../../interfaces/commands/flags/discord-command-flag";
 import { IDiscordCommandFlagError } from "../../../interfaces/commands/flags/discord-command-flag-error";
+import { IDiscordMessageFlag } from "../../../types/commands/flags/discord-message-flag";
 
 /**
  * @description
@@ -88,9 +89,9 @@ export abstract class DiscordCommandFlag<T extends string> {
     return example;
   }
 
-  public abstract isValid(messageFlag: Readonly<string>): boolean;
+  public abstract isValid(messageFlag: Readonly<IDiscordMessageFlag>): boolean;
 
   public abstract getInvalidFlagError(
-    messageFlag: Readonly<string>
+    messageFlag: Readonly<IDiscordMessageFlag>
   ): IDiscordCommandFlagError | null;
 }
