+++ src/features/discord/messages/functions/commands/flags/discord-command-get-flag-name.ts
@@ -1,4 +1,5 @@
 import _ from "lodash";
+import { IDiscordMessageFlag } from "../../../types/commands/flags/discord-message-flag";
 import { discordCommandRemoveFlagPrefix } from "./discord-command-remove-flag-prefix";
 
 /**
@@ -14,13 +15,13 @@ import { discordCommandRemoveFlagPrefix } from "./discord-command-remove-flag-pr
  * discordCommandGetFlagName('--f=')      => f
  * discordCommandGetFlagName('--f=dummy') => f
  *
- * @param {Readonly<string>} messageFlag A flag as a message
+ * @param {Readonly<IDiscordMessageFlag>} messageFlag A flag as a message
  * @param {Readonly<boolean>} toLowerCase Return the flag name to lower case
  *
  * @return {string | null} a string when the flag name exists
  */
 export function discordCommandGetFlagName(
-  messageFlag: Readonly<string>,
+  messageFlag: Readonly<IDiscordMessageFlag>,
   toLowerCase: Readonly<boolean> = false
 ): string | null {
   const flagName: string | undefined = _.head(
