+++ src/features/discord/messages/functions/commands/flags/discord-command-remove-flag-prefix.ts
@@ -1,4 +1,5 @@
 import _ from "lodash";
+import { IDiscordMessageFlag } from "../../../types/commands/flags/discord-message-flag";
 
 /**
  * @description
@@ -10,12 +11,12 @@ import _ from "lodash";
  * discordCommandRemoveFlagPrefix('-f')    => f
  * discordCommandRemoveFlagPrefix('--f')   => f
  *
- * @param {Readonly<string>} messageFlag A flag as a message
+ * @param {Readonly<IDiscordMessageFlag>} messageFlag A flag as a message
  *
  * @return {string} The flag message without the flag prefix "-" or "--"
  */
 export function discordCommandRemoveFlagPrefix(
-  messageFlag: Readonly<string>
+  messageFlag: Readonly<IDiscordMessageFlag>
 ): string {
   return _.trimStart(messageFlag, `--`);
 }
