+++ src/features/discord/messages/types/commands/flags/discord-message-flag.ts
@@ -1,0 +1,11 @@
+/**
+ * @description
+ * Is a text message containing only the flags part
+ * Useful for some functions to parse and extract information from it
+ * Without handling a long text message (good performances and cleaning enhancement)
+ *
+ * @example
+ * -e
+ * --enabled=true
+ */
+export type IDiscordMessageFlag = string;
